reafactored page
Refactored the page to now have a dropdown of all problems solved,
created an interface for all problems to implement.  Not using angular's
dependency injection because it requires to the component to have too
much in-depth knowledge about the underlying implementation (requiring
providers for the services doing the real work).  Perhaps these need to
be rolled up into some math module later.